requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
jinja2==3.1.2
httpx==0.25.2
openai==1.3.7
psycopg2-binary==2.9.9
tweepy==4.14.0
solders>=0.19.0
base58>=2.1.1
PyNaCl>=1.5.0
python-multipart